refactor: 클래스 책임 분리(분리기와 계산기)
- 클래스의 책임을 분리하기 위해 Calculator 클래스를 Separator와 Calculator 클래스로 분리
