+++ src/main/java/calculator/Application.java
@@ -4,9 +4,12 @@ public class Application {
     public static void main(String[] args) {
         ConsoleReader consoleReader = new ConsoleReader();
         ConsoleWriter consoleWriter = new ConsoleWriter();
+        Separator separator = new Separator();
         Calculator calculator = new Calculator();
+
         String input = consoleReader.readCalcLine();
-        int result = calculator.calc(input);
+        int[] nums = separator.separateInput(input);
+        int result = calculator.calcSum(nums);
         consoleWriter.writeCalcResult(result);
     }
 }
